requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-storages==1.8
entrypoints==0.3
flake8==3.7.9
idna==2.8
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.7.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.7
